requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-groq
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf
python-dotenv
tiktoken
beautifulsoup4
lxml
